third_party/blink/web_tests/external/wpt/html/canvas/element/the-canvas-state/2d.state.saverestore.shadowColor.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.state.saverestore.shadowColor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">

<h1>2d.state.saverestore.shadowColor</h1>

<script>

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 300;
  canvas.height = 150;
  const ctx = canvas.getContext('2d');

  // Test that restore() undoes any modifications.
  var old = ctx.shadowColor;
  ctx.save();
  ctx.shadowColor = "#ff0000";
  ctx.restore();
  _assertSame(ctx.shadowColor, old, "ctx.shadowColor", "old");

  // Also test that save() doesn't modify the values.
  ctx.shadowColor = "#ff0000";
  old = ctx.shadowColor;
  // We're not interested in failures caused by get(set(x)) != x (e.g.
  // from rounding), so compare against `old` instead of against "#ff0000".
  ctx.save();
  _assertSame(ctx.shadowColor, old, "ctx.shadowColor", "old");
  ctx.restore();
}, "save()/restore() works for shadowColor, with a canvas size of (300, 150)");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 0;
  canvas.height = 0;
  const ctx = canvas.getContext('2d');

  // Test that restore() undoes any modifications.
  var old = ctx.shadowColor;
  ctx.save();
  ctx.shadowColor = "#ff0000";
  ctx.restore();
  _assertSame(ctx.shadowColor, old, "ctx.shadowColor", "old");

  // Also test that save() doesn't modify the values.
  ctx.shadowColor = "#ff0000";
  old = ctx.shadowColor;
  // We're not interested in failures caused by get(set(x)) != x (e.g.
  // from rounding), so compare against `old` instead of against "#ff0000".
  ctx.save();
  _assertSame(ctx.shadowColor, old, "ctx.shadowColor", "old");
  ctx.restore();
}, "save()/restore() works for shadowColor, with a canvas size of (0, 0)");

</script>
